requirements.txt
pandas>=2.0
streamlit>=1.28
openpyxl>=3.1.0
fpdf2>=2.7.9
qrcode[pil]>=7.4.2
